+++ crates/stackable-operator/CHANGELOG.md
@@ -4,11 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
-- Support specifying externalTrafficPolicy in Services created by listener-operator ([#773], [#789]).
-- BREAKING: Rename `commons::listener::ServiceType` to `commons::listener::KubernetesServiceType` ([#773]).
+- Support specifying externalTrafficPolicy in Services created by listener-operator ([#773], [#789], [#791]).
 
 [#773]: https://github.com/stackabletech/operator-rs/pull/773
 [#789]: https://github.com/stackabletech/operator-rs/pull/789
+[#791]: https://github.com/stackabletech/operator-rs/pull/791
 
 ## [0.67.1] - 2024-05-08